+++ sdgym/synthesizers/base.py
@@ -51,8 +51,8 @@ class BaselineSynthesizer(abc.ABC):
             obj:
                 The synthesizer object.
         """
-        metadata_modality = MultiTableMetadata() if 'tables' in metadata else SingleTableMetadata()
-        metadata = metadata_modality.load_from_dict(metadata)
+        metadata_class = MultiTableMetadata() if 'tables' in metadata else SingleTableMetadata()
+        metadata = metadata_class.load_from_dict(metadata)
         return self._get_trained_synthesizer(data, metadata)
 
     def sample_from_synthesizer(self, synthesizer, n_samples):
